FASTAPI/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt==4.0.1
passlib[bcrypt]
python-jose
python-multipart
npm
google-generativeai
httpx
gunicorn
python-dotenv==0.19.2